Refactor TypeScript DTO parser
